# labels: optimize String method