Applying a pattern I learned about recently - `Write(AppendQuote(AvailableBuffer` does not allocate or copy when the buffer has sufficient space.

Use a stack buffer to reduce memory allocations.

Also some refactoring to add a benchmark without adding lines of code.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i7-14700K
         │  before.txt   │             after.txt              │
         │    sec/op     │   sec/op     vs base               │
String-4   1039.0n ± 29%   691.5n ± 7%  -33.45% (p=0.002 n=6)

         │ before.txt │             after.txt             │
         │    B/op    │    B/op     vs base               │
String-4   992.0 ± 0%   240.0 ± 0%  -75.81% (p=0.002 n=6)

         │ before.txt  │             after.txt             │
         │  allocs/op  │ allocs/op   vs base               │
String-4   16.000 ± 0%   1.000 ± 0%  -93.75% (p=0.002 n=6)
```